Update ParsePlugin.java
The Parse initialise command was returning success immediately, which would trigger the success callback.  This could mean that the Parse Installation record may not have saved to the Cloud in time.  This was no good for my needs as I want to use parse straight after the success callback is called.  This change will wait to call the success callback when the Installation has successfuly saved to the Cloud.  Or call the error callback if it can't save, eg. no internet connection.